pyproject.toml
[project]
name = "war-room-copilot"
version = "0.1.0"
description = "Voice-first AI agent for production incident war rooms"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[openai]>=1.0",
    "livekit-plugins-silero",
    "python-dotenv",
    "pydantic>=2.0",
]

[project.optional-dependencies]
all = [
    "livekit-plugins-speechmatics",
    "fastapi",
    "uvicorn[standard]",
    "websockets",
    "aiosqlite",
    "pygithub",
    "langsmith",
    "httpx",
    "pyyaml",
]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "pytest-asyncio",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[tool.mypy]
python_version = "3.12"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
